+++ experimental/packages/opentelemetry-configuration/src/EnvironmentConfigProvider.ts
@@ -26,6 +26,11 @@ import {
   getNumberFromEnv,
 } from '@opentelemetry/core';
 import { ConfigProvider } from './IConfigProvider';
+import {
+  ExemplarFilter,
+  ExporterDefaultHistogramAggregation,
+  ExporterTemporalityPreference,
+} from './models/meterProviderModel';
 
 /**
  * EnvironmentConfigProvider provides a configuration based on environment variables.
@@ -257,6 +262,9 @@ function setMeterProvider(config: ConfigurationModel): void {
     if (timeout) {
       readerPeriodic.timeout = timeout;
     }
+    if (readerPeriodic.exporter.otlp_http == null) {
+      readerPeriodic.exporter.otlp_http = {};
+    }
 
     const endpoint = getStringFromEnv('OTEL_EXPORTER_OTLP_METRICS_ENDPOINT');
     if (endpoint) {
@@ -305,26 +313,45 @@ function setMeterProvider(config: ConfigurationModel): void {
     const temporalityPreference = getStringFromEnv(
       'OTEL_EXPORTER_OTLP_METRICS_TEMPORALITY_PREFERENCE'
     );
-    if (
-      temporalityPreference &&
-      (temporalityPreference === 'cumulative' ||
-        temporalityPreference === 'delta' ||
-        temporalityPreference === 'low_memory')
-    ) {
-      readerPeriodic.exporter.otlp_http.temporality_preference =
-        temporalityPreference;
+    if (temporalityPreference) {
+      switch (temporalityPreference) {
+        case 'cumulative':
+          readerPeriodic.exporter.otlp_http.temporality_preference =
+            ExporterTemporalityPreference.Cumulative;
+          break;
+        case 'delta':
+          readerPeriodic.exporter.otlp_http.temporality_preference =
+            ExporterTemporalityPreference.Delta;
+          break;
+        case 'low_memory':
+          readerPeriodic.exporter.otlp_http.temporality_preference =
+            ExporterTemporalityPreference.LowMemory;
+          break;
+        default:
+          readerPeriodic.exporter.otlp_http.temporality_preference =
+            ExporterTemporalityPreference.Cumulative;
+          break;
+      }
     }
 
     const defaultHistogramAggregation = getStringFromEnv(
       'OTEL_EXPORTER_OTLP_METRICS_DEFAULT_HISTOGRAM_AGGREGATION'
     );
-    if (
-      defaultHistogramAggregation &&
-      (defaultHistogramAggregation === 'explicit_bucket_histogram' ||
-        defaultHistogramAggregation === 'base2_exponential_bucket_histogram')
-    ) {
-      readerPeriodic.exporter.otlp_http.default_histogram_aggregation =
-        defaultHistogramAggregation;
+    if (defaultHistogramAggregation) {
+      switch (defaultHistogramAggregation) {
+        case 'explicit_bucket_histogram':
+          readerPeriodic.exporter.otlp_http.default_histogram_aggregation =
+            ExporterDefaultHistogramAggregation.ExplicitBucketHistogram;
+          break;
+        case 'base2_exponential_bucket_histogram':
+          readerPeriodic.exporter.otlp_http.default_histogram_aggregation =
+            ExporterDefaultHistogramAggregation.Base2ExponentialBucketHistogram;
+          break;
+        default:
+          readerPeriodic.exporter.otlp_http.default_histogram_aggregation =
+            ExporterDefaultHistogramAggregation.ExplicitBucketHistogram;
+          break;
+      }
     }
     if (config.meter_provider == null) {
       config.meter_provider = { readers: [{}] };
@@ -336,16 +363,24 @@ function setMeterProvider(config: ConfigurationModel): void {
     config.meter_provider.readers[0].periodic = readerPeriodic;
   }
   const exemplarFilter = getStringFromEnv('OTEL_METRICS_EXEMPLAR_FILTER');
-  if (
-    exemplarFilter &&
-    (exemplarFilter === 'trace_based' ||
-      exemplarFilter === 'always_on' ||
-      exemplarFilter === 'always_off')
-  ) {
+  if (exemplarFilter) {
     if (config.meter_provider == null) {
-      config.meter_provider = {};
+      config.meter_provider = { readers: [] };
     }
-    config.meter_provider.exemplar_filter = exemplarFilter;
+    switch (exemplarFilter) {
+      case 'trace_based':
+        config.meter_provider.exemplar_filter = ExemplarFilter.TraceBased;
+        break;
+      case 'always_on':
+        config.meter_provider.exemplar_filter = ExemplarFilter.AlwaysOn;
+        break;
+      case 'always_off':
+        config.meter_provider.exemplar_filter = ExemplarFilter.AlwaysOff;
+        break;
+      default:
+        config.meter_provider.exemplar_filter = ExemplarFilter.TraceBased;
+        break;
+    }
   }
 }
 
@@ -358,7 +393,7 @@ function setLoggerProvider(config: ConfigurationModel): void {
   );
   if (attributeValueLengthLimit || attributeCountLimit) {
     if (config.logger_provider == null) {
-      config.logger_provider = {};
+      config.logger_provider = { processors: [] };
     }
     if (config.logger_provider.limits == null) {
       config.logger_provider.limits = { attribute_count_limit: 128 };